Extract mock pkg
